move schema and fix query
